Consolidate top-level styles, remove obsolete ones
