--- src/styled/developer-topics.ts
@@ -1,15 +1,0 @@
-import styled from '@emotion/styled';
-
-const StyledButton = styled('div')`
-    background: none;
-    border: none;
-    cursor: pointer;
-`;
-
-const DeveloperTopicsTab = styled('div')`
-    align-items: center;
-    display: flex;
-    justify-content: center;
-`;
-
-export { StyledButton, DeveloperTopicsTab };
