--- src/styled/styles.ts
@@ -1,5 +1,0 @@
-export const pageWrapper = {
-    paddingBottom: 'inc160',
-    px: ['inc40', null, 'inc50', 'inc70'],
-    paddingTop: ['inc40', null, 'inc50', 'inc70'],
-};
